New module on Artificial Neural Networks added. Changes include:
1. New folder 'ANN' created in include and src
2. A new class TANN declared in include/ANN/ANN.h and defined in
src/ANN/ANN.C
3. CMakeLists.txt file added in src/ANN to compile the ANN relevant
files and create a static library 'ann'
4. CMakeLists.txt file of the main code in the root dir. modified to
include the ANN Folders as well as the static library ann

This new class 'TANN' would contain the neural network, read and write
functionality as well as  metadata. This class in turn invokes functions
from MLPACK. All the MLPACK includes are ONLY declared in ANN.h and
nowhere else in ParMooN code to keep it clean.

To use this class, just include TANN.h in ParMooN code and start using
the class as well as other MLPACK features.