First working ANN code in ParMooN
The changes include:

1. ANN.h contains a template class TANN which is completely declared and
defined in ANN.h. This is the central class to process the neural
network.

2. A new class called TANNDatasetHandler added. This class stores the
dataset, training and testing labels and errors. A pointer to the object
of this class is passed to TANN class to train the network as well as to
predict values.

3. ANNParamReader completely revamped. It now uses mostly C++ functions
to read the lines of the param file and create the required arrays.

4. Many minor changes and revision in almost all of the files as a
result of debugging and cleaning up.

This is the first working version of the code. Some more cleaning,
optimization and value addition is expected in the coming days / future
commits.